+++ pmdictionaryvalidation/src/test/java/org/onap/validation/cli/command/validate/ValidateYamlCommandTest.java
@@ -29,7 +29,7 @@ import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.onap.validation.cli.util.ToResponseModelConverter.toModel;
+import static org.onap.validation.cli.command.validate.ToResponseModelConverter.toModel;
 
 @ExtendWith(MockitoExtension.class)
 class ValidateYamlCommandTest {